requirements.txt
django
djangorestframework
pyforkurento